requirements.txt
asgiref==3.2.7
bson==0.5.8
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
dataclasses==0.6
Django==3.0.7
django-cors-headers==3.2.1
django-ipware==2.1.0
django-user-agents==0.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djecrety==1.0.13
favicon==0.7.0
httpie==2.1.0
idna==2.9
jsonfield==3.1.0
Pillow==7.2.0
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0.1
sqlparse==0.2.4
ua-parser==0.10.0
urllib3==1.25.9
user-agents==2.1
